Move internal catalog cache loading into tables service
This keeps the cache manager owned by the tables app so metadata cache settings bind through the shared Spring environment without creating the cache in other services.

Made-with: Cursor
