+++ iceberg/openhouse/internalcatalog/build.gradle
@@ -19,8 +19,6 @@ dependencies {
   api project(':cluster:metrics')
   api project(':iceberg:azure')
   api project(':services:common')
-  implementation 'org.springframework:spring-context-support:5.3.18'
-  implementation 'com.github.ben-manes.caffeine:caffeine:2.8.8'
   // Needed for testing against manually bootstrapped HTS server through bootRun task.
   testImplementation "io.netty:netty-resolver-dns-native-macos:4.1.70.Final:osx-x86_64"
   // Needed for testing against mock HTS server along with other constructs in /tables module.
