+++ services/tables/build.gradle
@@ -40,6 +40,8 @@ dependencies {
   api project(':cluster:metrics')
   api 'org.springframework.security:spring-security-config:5.7.2'
   api 'org.springframework.boot:spring-boot-starter-webflux:2.7.8'
+  implementation 'org.springframework:spring-context-support:5.3.18'
+  implementation 'com.github.ben-manes.caffeine:caffeine:2.8.8'
   implementation 'com.cronutils:cron-utils:9.2.0'
   testImplementation 'org.junit.jupiter:junit-jupiter-engine:' + junit_version
   testImplementation 'org.springframework.security:spring-security-test:5.7.3'
